Файлы библиотеки CubeF4 обновлены до 28.1

+++ sources/common/libs/STM32Cube_F4/Middlewares/ST/STM32_USB_Device_Library/Core/Inc/usbd_def.h
@@ -432,7 +432,7 @@ __STATIC_INLINE uint16_t SWAPBYTE(uint8_t *addr)
   _pbuff++;
   _Byte2 = *(uint8_t *)_pbuff;
 
-  _SwapVal = (uint16_t)((_Byte2 << 8) | _Byte1);
+  _SwapVal = (_Byte2 << 8) | _Byte1;
 
   return _SwapVal;
 }
